+++ dwds/test/integration/instances/patterns_inspection_ddc_library_bundle_test.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2024, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
@@ -7,19 +7,19 @@
 @Timeout(Duration(minutes: 2))
 library;
 
-import 'package:dwds/src/services/expression_compiler.dart';
+import 'package:dwds/expression_compiler.dart';
 import 'package:test/test.dart';
 import 'package:test_common/test_sdk_configuration.dart';
 
 import '../fixtures/context.dart';
-import 'common/record_inspection_common.dart';
+import 'common/patterns_inspection_common.dart';
 
 void main() {
   // Enable verbose logging for debugging.
   const debug = false;
-  final canaryFeatures = true;
 
   group('canary: true | Frontend Server |', () {
+    final canaryFeatures = true;
     final compilationMode = CompilationMode.frontendServer;
     final provider = TestSdkConfigurationProvider(
       verbose: debug,
@@ -27,6 +27,7 @@ void main() {
       ddcModuleFormat: ModuleFormat.ddc,
     );
     tearDownAll(provider.dispose);
+
     runTests(
       provider: provider,
       compilationMode: compilationMode,
@@ -36,6 +37,7 @@ void main() {
   });
 
   group('canary: true | Build Daemon |', () {
+    final canaryFeatures = true;
     final compilationMode = CompilationMode.buildDaemon;
     final provider = TestSdkConfigurationProvider(
       verbose: debug,
@@ -43,6 +45,7 @@ void main() {
       ddcModuleFormat: ModuleFormat.ddc,
     );
     tearDownAll(provider.dispose);
+
     runTests(
       provider: provider,
       compilationMode: compilationMode,