+++ dwds/test/integration/instances/record_type_inspection_ddc_library_bundle_test.dart
@@ -18,9 +18,25 @@ void main() {
   // Enable verbose logging for debugging.
   const debug = false;
   final canaryFeatures = true;
-  final compilationMode = CompilationMode.frontendServer;
 
-  group('canary: $canaryFeatures |', () {
+  group('canary: true | Frontend Server |', () {
+    final compilationMode = CompilationMode.frontendServer;
+    final provider = TestSdkConfigurationProvider(
+      verbose: debug,
+      canaryFeatures: canaryFeatures,
+      ddcModuleFormat: ModuleFormat.ddc,
+    );
+    tearDownAll(provider.dispose);
+    runTests(
+      provider: provider,
+      compilationMode: compilationMode,
+      canaryFeatures: canaryFeatures,
+      debug: debug,
+    );
+  });
+
+  group('canary: true | Build Daemon |', () {
+    final compilationMode = CompilationMode.buildDaemon;
     final provider = TestSdkConfigurationProvider(
       verbose: debug,
       canaryFeatures: canaryFeatures,